requirements.txt
fastapi
uvicorn
pandas
sentence-transformers
numpy
faiss-gpu